+++ app/Controllers/Admin/JobController.php
@@ -30,7 +30,7 @@ class JobController extends Controller
     {
         Job::create(request()->all());
 
-        return redirect('/admin/jobs');
+        return redirect('/admin/jobs', compact(flash()->add('Thêm thành công', 'success')));
     }
 
     public function edit()
@@ -48,7 +48,7 @@ class JobController extends Controller
 
         $job->update(request()->all());
 
-        return redirect('/admin/jobs');
+        return redirect('/admin/jobs', compact(flash()->add('Sửa thành công', 'success')));
     }
 
     public function destroy()
@@ -57,6 +57,6 @@ class JobController extends Controller
 
         $job->delete();
 
-        return redirect('/admin/jobs');
+        return redirect('/admin/jobs', compact(flash()->add('Xóa thành công', 'success')));
     }
 }
